+++ Uniceps/Commands/RoutineSystemCommands/RoutineItemsCommands/DeleteRoutineItemCommand.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using Uniceps.Commands;
 using Uniceps.Stores.RoutineStores;
 
@@ -19,7 +20,11 @@ namespace Uniceps.Commands.RoutineSystemCommands.RoutineItemsCommands
 
         public override async Task ExecuteAsync(object? parameter)
         {
-            await _routineItemDataStore.Delete(_routineItemDataStore.SelectedRoutineItem!.Id);
+            if(MessageBox.Show("سيتم حذف التمرين ... هل انت متاكد","",MessageBoxButton.YesNo,MessageBoxImage.Warning)== MessageBoxResult.Yes)
+            {
+                await _routineItemDataStore.Delete(_routineItemDataStore.SelectedRoutineItem!.Id);
+
+            }
         }
     }
 }
